Move resource management out of CommCarePlatform
