+++ application/src/org/commcare/applogic/CommCareUpgradeState.java
@@ -12,6 +12,7 @@ import org.commcare.resources.model.UnresolvedResourceException;
 import org.commcare.util.CommCareContext;
 import org.commcare.util.CommCareInitializer;
 import org.commcare.util.CommCarePlatform;
+import org.commcare.util.CommCareResourceManager;
 import org.commcare.util.InitializationListener;
 import org.commcare.util.YesNoListener;
 import org.commcare.view.CommCareStartupInteraction;
@@ -65,7 +66,6 @@ public abstract class CommCareUpgradeState implements State, TrivialTransitions 
             ResourceTable recovery = CommCareContext.CreateTemporaryResourceTable(RECOVERY_TABLE_NAME);
 
             protected boolean runWrapper() throws UnfullfilledRequirementsException {
-
                 if(networkRetries != -1) {
                     upgrade.setNumberOfRetries(networkRetries);
                 }
@@ -80,7 +80,9 @@ public abstract class CommCareUpgradeState implements State, TrivialTransitions 
 
                 while(!staged) {
                     try {
-                        CommCareContext._().getManager().stageUpgradeTable(CommCareContext.RetrieveGlobalResourceTable(), upgrade, recovery, false);
+                        CommCareResourceManager resourceManager = 
+                            new CommCareResourceManager(CommCareContext._().getManager());
+                        resourceManager.stageUpgradeTable(CommCareContext.RetrieveGlobalResourceTable(), upgrade, recovery, false);
                         interaction.updateProgess(20);
                         staged = true;
                     } catch (UnresolvedResourceException e) {
@@ -126,7 +128,7 @@ public abstract class CommCareUpgradeState implements State, TrivialTransitions 
                     public void resourceStateUpdated(ResourceTable table) {
                         int score = 0;
                         int max = 0;
-                        Vector<Resource> resources = CommCarePlatform.getResourceListFromProfile(table);
+                        Vector<Resource> resources = CommCareResourceManager.getResourceListFromProfile(table);
                         max = resources.size() * INSTALL_SCORE;
 
                         if(max <= INSTALL_SCORE*2) {
@@ -182,8 +184,10 @@ public abstract class CommCareUpgradeState implements State, TrivialTransitions 
                     upgradeAttemptPending = false;
 
                     try {
-                        CommCareContext._().getManager().prepareUpgradeResources(global, upgrade, recovery);
-                        CommCareContext._().getManager().upgrade(global, upgrade, recovery);
+                        CommCareResourceManager resourceManager = 
+                            new CommCareResourceManager(CommCareContext._().getManager());
+                        resourceManager.prepareUpgradeResources(global, upgrade, recovery);
+                        resourceManager.upgrade(global, upgrade, recovery);
                     } catch(UnreliableSourceException e) {
                         //We simply can't retrieve all of the resources that we're looking for.
 
@@ -224,7 +228,7 @@ public abstract class CommCareUpgradeState implements State, TrivialTransitions 
                 String logMsg = "Upgrade attempt unsuccesful. Probably due to network. ";
 
                 //Count resources
-                Vector<Resource> resources = CommCarePlatform.getResourceListFromProfile(upgrade);
+                Vector<Resource> resources = CommCareResourceManager.getResourceListFromProfile(upgrade);
                 int downloaded = 0;
 
                 for(Resource r : resources ){
